Split App.java into separate classes and packages
