+++ src/main/java/org/example/service/ApiClient.java
@@ -1,0 +1,29 @@
+package org.example.service;
+
+import java.net.URI;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
+
+public class ApiClient {
+    public static HttpResponse<String> getPrices(String url_string) {
+        HttpResponse<String> response = null;
+        try {
+            HttpClient client = HttpClient.newHttpClient();
+            HttpRequest request = HttpRequest.newBuilder()
+                    .uri(URI.create(url_string))
+                    .build();
+
+        response = client.send(request, HttpResponse.BodyHandlers.ofString());
+
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return response;
+    }
+
+    public static String urlBuilder(String date, String zone) {
+        String year = String.valueOf(java.time.LocalDate.now().getYear());
+        return "https://www.elprisetjustnu.se/api/v1/prices/" + year + "/" + date + "_" + zone + ".json";
+    }
+}
